+++ spartacus/src/angle_conversion_callbacks.py
@@ -1,7 +1,8 @@
 import biorbd
 import numpy as np
 from .enums import EulerSequence
-from .utils import BiomechCoordinateSystem, mat_2_rotation
+from .utils import mat_2_rotation
+from .biomech_system import BiomechCoordinateSystem
 
 
 def get_angle_conversion_callback_from_tuple(tuple_factors: tuple[int, int, int]) -> callable:
